refactor: rename views/widgets/ → views/components/
After Task 2-5 cleanup the directory only holds DOM-producing
reusable elements. "Components" describes them more accurately than
"widgets". Pure rename + import path rewrite.
