+++ app/src/scene/components/buildings/iconAtlas.ts
@@ -21,7 +21,7 @@
 import * as THREE from 'three';
 import { NodeKind } from '@/types';
 import type { DirNode, FileNode, Manifest, TreeNode } from '@/types';
-import { FILE_ICON_CDN_BASE, getFileIconName } from '@/views/widgets/fileIcon.js';
+import { FILE_ICON_CDN_BASE, getFileIconName } from '@/views/components/fileIcon.js';
 
 // Atlas is 2048×2048 with 128-px slots → up to 16×16 = 256 unique
 // icons at 4× the per-icon resolution of the old 64-px slots.
