+++ app/src/styles.css
@@ -741,7 +741,7 @@ h6 {
    instead of being collapsed to currentColor. Same 1em sizing as
    .lucide-icon so it slots into header buttons without throwing off
    alignment. The .lucide-icon mask path is the simple monochrome variant;
-   see makeGemIcon() in views/widgets/icon.ts. */
+   see makeGemIcon() in views/components/icon.ts. */
 .gem-icon {
   display: inline-block;
   width: 1em;
