+++ app/src/views/components/fileIcon.ts
@@ -1,4 +1,4 @@
-// views/widgets/fileIcon.ts — VSCode-style file/folder icons (Material
+// views/components/fileIcon.ts — VSCode-style file/folder icons (Material
 // Icon Theme, MIT). Returns an <img> element pointing at a CDN-hosted
 // SVG; the browser caches per URL so a project with N files but only K
 // unique extensions only triggers K icon fetches on first paint.
