+++ app/src/views/shell/appHeader.ts
@@ -7,10 +7,10 @@
 // When no path is selected (or only the root), #app-title is empty.
 
 import { ASPHALT, BUILDING_PALETTE } from '@/config';
-import { makeGemIcon, makeLucideIcon } from '../widgets/icon.js';
-import { makeExtensionBadge } from '../widgets/badge.js';
+import { makeGemIcon, makeLucideIcon } from '../components/icon.js';
+import { makeExtensionBadge } from '../components/badge.js';
 import { toHttpsRepoUrl } from '@/utils/sources.js';
-import { fitSegments } from '../widgets/pathTruncate.js';
+import { fitSegments } from '../components/pathTruncate.js';
 
 // How long the "Copied!" badge lingers after the copy button is clicked.
 const COPY_FEEDBACK_DURATION_MS = 1500;
