+++ app/tests/views/components/sourcePicker.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, beforeEach, vi } from 'vitest';
-import { createSourcePicker } from '@/views/widgets/sourcePicker.js';
+import { createSourcePicker } from '@/views/components/sourcePicker.js';
 import { pushRecent } from '@/utils/sourceRecents.js';
 
 function mountRoot(): HTMLElement {
