+++ app/tests/views/shell/tooltip.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, afterEach } from 'vitest';
-import { showTooltip, hideTooltip, moveTooltip } from '@/views/widgets/tooltip.js';
+import { showTooltip, hideTooltip, moveTooltip } from '@/views/components/tooltip.js';
 
 describe('tooltip', () => {
   afterEach(() => {
